# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Household task management API.\n\nConventions callers need:\n- Every response is the envelope `{ success, message, data }`.\n- `user`, `group` and `task` are addressed by opaque `uid`; every other resource by numeric `id`.\n- Timestamps are UTC with no zone suffix — parse as UTC, convert for display.\n- Actors are derived from the token; never send a user or member id to identify the caller.\n- Errors: 400 validation, 401 unauthenticated, 403 forbidden, 404 not found, 409 duplicate or violated business rule, 429 rate limited (see `Retry-After`).\n- Role requirements are stated per operation and are not modelled in this spec.\n",
    "title" : "TasKlean API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this API"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/tags" : {
      "get" : {
        "description" : "Lists a group's active tags.",
        "operationId" : "getTagsByGroup",
        "parameters" : [ {
          "description" : "the group id",
          "in" : "query",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTagResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the tags"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists a group's active tags.",
        "tags" : [ "Tags" ]
      },
      "post" : {
        "description" : "Creates a tag.",
        "operationId" : "createTag",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TagRequest"
              }
            }
          },
          "description" : "the tag details (name, color, group)",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTagResponse"
                }
              }
            },
            "description" : "<code>201 Created</code> with the created tag"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Creates a tag.",
        "tags" : [ "Tags" ]
      }
    },
    "/api/tags/{id}" : {
      "delete" : {
        "description" : "Soft-deletes a tag.",
        "operationId" : "deleteTag",
        "parameters" : [ {
          "description" : "the tag id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "<code>200 OK</code> with a confirmation message"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Soft-deletes a tag.",
        "tags" : [ "Tags" ]
      },
      "get" : {
        "description" : "Fetches a tag by id.",
        "operationId" : "getTag",
        "parameters" : [ {
          "description" : "the tag id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTagResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the tag"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Fetches a tag by id.",
        "tags" : [ "Tags" ]
      },
      "put" : {
        "description" : "Updates a tag.",
        "operationId" : "updateTag",
        "parameters" : [ {
          "description" : "the tag id",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TagRequest"
              }
            }
          },
          "description" : "the new tag details",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTagResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the updated tag"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Updates a tag.",
        "tags" : [ "Tags" ]
      }
    },
...
    "/api/tasks" : {
      "get" : {
        "description" : "Lists a group's active tasks.",
        "operationId" : "getTasksByGroup",
        "parameters" : [ {
          "description" : "the group id",
          "in" : "query",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTaskResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the tasks"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists a group's active tasks.",
        "tags" : [ "Tasks" ]
      },
      "post" : {
        "description" : "Creates a task.",
        "operationId" : "createTask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskRequest"
              }
            }
          },
          "description" : "the task details",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "<code>201 Created</code> with the created task"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Creates a task.",
        "tags" : [ "Tasks" ]
      }
    },
    "/api/tasks/{uid}" : {
      "delete" : {
        "description" : "Soft-deletes a task.",
        "operationId" : "deleteTask",
        "parameters" : [ {
          "description" : "the task's public UID",
          "in" : "path",
          "name" : "uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "<code>200 OK</code> with a confirmation message"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Soft-deletes a task.",
        "tags" : [ "Tasks" ]
      },
      "get" : {
        "description" : "Fetches a task by public UID.",
        "operationId" : "getTask",
        "parameters" : [ {
          "description" : "the task's public UID",
          "in" : "path",
          "name" : "uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the task"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Fetches a task by public UID.",
        "tags" : [ "Tasks" ]
      },
      "put" : {
        "description" : "Updates a task.",
        "operationId" : "updateTask",
        "parameters" : [ {
          "description" : "the task's public UID",
          "in" : "path",
          "name" : "uid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskRequest"
              }
            }
          },
          "description" : "the new task details",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the updated task"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Updates a task.",
        "tags" : [ "Tasks" ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Validation failure, malformed body, or a bad path/query parameter"
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Authenticated, but lacking the platform or group role this operation requires"
      },
      "InternalServerError" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Unexpected server error"
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Rate limit exceeded — see the Retry-After header"
      },
      "Unauthorized" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "No valid access token — or, on the public auth endpoints, rejected credentials"
      }
    },
    "schemas" : {
...
      "ApiResponseListTagResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TagResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseListTaskResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TaskResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseTagResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TagResponse"
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseTaskResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TaskResponse"
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseVoid" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : { },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "description" : "The standard envelope, as returned for a failed request.",
        "properties" : {
          "data" : {
            "type" : "object",
            "description" : "Always null on an error"
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message"
          },
          "success" : {
            "type" : "boolean",
            "default" : false
          }
        }
      },
...
      "TagRequest" : {
        "type" : "object",
        "properties" : {
          "color" : {
            "type" : "string"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "groupId", "name" ]
      },
      "TagResponse" : {
        "type" : "object",
        "properties" : {
          "color" : {
            "type" : "string"
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateUpdated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
      "TaskRequest" : {
        "type" : "object",
        "properties" : {
          "assignedToId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "categoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdById" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "estimatedTimeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "nextDueDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "priority" : {
            "type" : "string",
            "minLength" : 1
          },
          "recurrencePattern" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "recurrenceType" : {
            "type" : "string"
          },
          "requiresPhotoProof" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string",
            "minLength" : 1
          },
          "taskPhotoUrl" : {
            "type" : "string"
          }
        },
        "required" : [ "createdById", "groupId", "name", "priority", "status" ]
      },
      "TaskResponse" : {
        "type" : "object",
        "properties" : {
          "assignedToId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "categoryId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "createdById" : {
            "type" : "integer",
            "format" : "int64"
          },
          "dateCompleted" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateUpdated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "description" : {
            "type" : "string"
          },
          "estimatedTimeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "groupId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "nextDueDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "priority" : {
            "type" : "string"
          },
          "recurrencePattern" : {
            "type" : "object",
            "additionalProperties" : { }
          },
          "recurrenceType" : {
            "type" : "string"
          },
          "requiresPhotoProof" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          },
          "taskPhotoUrl" : {
            "type" : "string"
          },
          "uid" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/auth/login, /api/auth/verify-email, /api/auth/google or /api/auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}